SED-3706: reusable getTestSetStatusOverview
